Extract report format and report output
